UpdatesOnUI
changed the UI canvas to overlay. changed some creature prefab order in layer. pulled out some backgrounds and the drawing canvas from the UI canvas. Updated inventory box scroll: now it has a feathered border and a container. Added a line to the inventory script so that it shows name. Updated setting and collection interaction.